Update cascade.less: brand text breaks on login page and all textes on Sidenav break when too long
Someone requested a change for too long hostnames/brand textes not breaking correctly.

This would be the easiest change. However the text which appears in the mobile version without the sidenav still does not break. I believe it should be deleted anyway for clearer design. Or make the text smaller in mobile view and put a wrapper around it which truncates the word with "..." at the end after 16 chars.